requirements.txt
asgiref==3.11.0
apscheduler==3.10.4
black==25.11.0
channels==4.3.2
click==8.3.1
Django==6.0
django-environ==0.12.0
djangorestframework==3.16.1
drf-yasg==1.21.11
flake8==7.3.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.3.0
isort==7.0.0
joblib==1.5.2
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
psycopg==3.3.1
psycopg-binary==3.3.1
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.4
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.11.0
